Fix recursive_mutex reentrancy check truncating the thread id to 32 bits
_Try_wait() compared m_tid (uint64_t, holding get_current_thread_id()'s
full pthread_self()-derived value on POSIX) against a local uint32_t
copy of the same call, silently truncating it. pthread_t is typically
a pointer-sized value on Linux/macOS, so the truncated comparison
almost never matched even for the owning thread -- reentrancy from the
constructing thread failed reliably on every non-Windows CI runner
(GetCurrentThreadId() happens to fit in 32 bits, which is why this was
invisible locally on MSVC).

+++ mst/mx_threading_slim.inl
@@ -431,7 +431,7 @@ inline void recursive_mutex::signal() const noexcept
 
 inline bool recursive_mutex::_Try_wait() const noexcept
 {
-	const uint32_t tid = ::mst::_Details::get_current_thread_id();
+	const uint64_t tid = ::mst::_Details::get_current_thread_id();
 
 	if(m_tid == tid)
 	{
